refactor: add a super method chatMessages to convert dicts to model
